% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ess.R
\name{compute_ess}
\alias{compute_ess}
\title{Computing the effective sample size}
\usage{
compute_ess(gamma_current, gamma_proposal, logweights_IS)
}
\arguments{
\item{gamma_current}{the current inverse temperature}

\item{gamma_proposal}{the proposed inverse temperature}

\item{logweights_IS}{the log ratio of the target density over the proposal density}
}
\description{
Computing the effective sample size
}
