% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_kernel.R
\name{construct_kernel}
\alias{construct_kernel}
\title{Construct the MCMC rejuvenation kernel}
\usage{
construct_kernel(target, proposal, mcmc_kernel, gamma_current)
}
\arguments{
\item{target}{list with keys: \code{logdensity} evaluates log target density, \code{gradlogdensity} returns its gradient}

\item{proposal}{list with keys: \code{rinit} samples from proposal, \code{logdensity} evaluates log proposal density, \code{gradlogdensity} returns its gradient}

\item{mcmc_kernel}{list with keys: \code{choice} MCMC method, \code{parameters} associated tuning parameters, \code{nmcmc} number of MCMC moves for each intermediate distribution}

\item{gamma_current}{the current inverse temperature}
}
\description{
Construct the MCMC rejuvenation kernel
}
