% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_next_temp.R
\name{find_next_temp}
\alias{find_next_temp}
\title{Binary search the next inverse temperature}
\usage{
find_next_temp(gamma_current, ess_criterion, logweights_IS)
}
\arguments{
\item{gamma_current}{the current inverse temperature}

\item{ess_criterion}{the designated effective sample size threshold}

\item{logweights_IS}{the log ratio of the target density over the proposal density}
}
\description{
Binary search the next inverse temperature
}
