% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mvnorm.R
\name{get_mvnormal}
\alias{get_mvnormal}
\title{Get multivariate normal target}
\usage{
get_mvnormal(dimension, mean_target, Sigma_target)
}
\arguments{
\item{dimension}{dimension of the target}

\item{mean_target}{mean}

\item{Sigma_target}{covariance}
}
\value{
list with keys \code{logdensity} and \code{gradlogdensity},
}
\description{
Define a multivariate normal target
with prescribed dimension, mean, covariance.
}
