% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_student_t.R
\name{get_student_t}
\alias{get_student_t}
\title{Get student t distribution}
\usage{
get_student_t(dimension, degree, mean_proposal, Sigma_proposal)
}
\arguments{
\item{dimension}{dimension of the target}

\item{degree}{degrees of freedom}

\item{mean_target}{mean}

\item{Sigma_target}{covariance}
}
\value{
list with keys \code{logdensity} and \code{gradlogdensity},
}
\description{
Define a student t distribution
with prescribed dimension, mean, variance.
}
