% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{gibbs}
\alias{gibbs}
\title{Run the Metropolis-within-Gibbs algorithm}
\usage{
gibbs(target, proposal, gibbs_parameters, gamma_current)
}
\arguments{
\item{target}{list with keys: \code{logdensity} evaluates log target density, \code{gradlogdensity} returns its gradient}

\item{proposal}{list with keys: \code{rinit} samples from proposal, \code{logdensity} evaluates log proposal density, \code{gradlogdensity} returns its gradient}

\item{gibbs_parameters}{a vector of step sizes used in the Metroplis algorithm}

\item{gamma_current}{the current inverse temperature}
}
\description{
Run the Metropolis-within-Gibbs algorithm
}
