% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc}
\alias{hmc}
\title{Run Hamiltonian Monte Carlo}
\usage{
hmc(target, proposal, hmc_parameters, gamma_current)
}
\arguments{
\item{target}{list with keys: \code{logdensity} evaluates log target density, \code{gradlogdensity} returns its gradient}

\item{proposal}{list with keys: \code{rinit} samples from proposal, \code{logdensity} evaluates log proposal density, \code{gradlogdensity} returns its gradient}

\item{hmc_parameters}{list with keys: \code{stepsize} leapfrog step size, \code{nsteps}, number of leapfrog steps}

\item{gamma_current}{the current inverse temperature}
}
\description{
Run Hamiltonian Monte Carlo
}
