% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_weight.R
\name{normalize_weight}
\alias{normalize_weight}
\title{Normalize vector of log-weights}
\usage{
normalize_weight(logweights)
}
\arguments{
\item{logweights}{is a vector of n real values}
}
\value{
a vector of n non-negative values summing to one
}
\description{
Takes a vector of real values, and return the vector of normalized exponentiated values,
as well as average of the exponentiated unnormalized values.
}
\examples{
N <- 1000
logweights <- rnorm(N)
normalize_weight_results <- normalize_weight(logweights)
normalized_weights <- normalize_weight_results$nw
}
