% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_hmc.R
\name{rm_hmc}
\alias{rm_hmc}
\title{Run Riemann manifold Hamiltonian Monte Carlo}
\usage{
rm_hmc(target, proposal, rm_hmc_parameters, gamma_current)
}
\arguments{
\item{target}{list with keys: \code{logdensity} evaluates log target density, \code{gradlogdensity} returns its gradient}

\item{proposal}{list with keys: \code{rinit} samples from proposal, \code{logdensity} evaluates log proposal density, \code{gradlogdensity} returns its gradient}

\item{rm_hmc_parameters}{list with keys: \code{stepsize} step size in the leap-frog integrator, \code{nsteps} number of leapfrog steps}

\item{gamma_current}{the current inverse temperature}
}
\description{
Run Riemann manifold Hamiltonian Monte Carlo
}
