% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{run_smc}
\alias{run_smc}
\title{Run sequential Monte Carlo sampler}
\usage{
run_smc(target, proposal, nparticles, mcmc_kernel, ess_criterion)
}
\arguments{
\item{target}{list with keys: \code{logdensity} evaluates log target density, \code{gradlogdensity} returns its gradient}

\item{proposal}{list with keys: \code{rinit} samples from proposal, \code{logdensity} evaluates log proposal density, \code{gradlogdensity} returns its gradient}

\item{nparticles}{number of particles}

\item{mcmc_kernel}{list with keys: \code{choice} MCMC method, \code{parameters} associated tuning parameters, \code{nmcmc} number of MCMC moves for each intermediate distribution}

\item{ess_criterion}{value between [0,1] which controls ESS adaptation}
}
\description{
Run sequential Monte Carlo sampler
}
