% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_gibbs.R
\name{slice_gibbs}
\alias{slice_gibbs}
\title{Run slice Gibbs sampler}
\usage{
slice_gibbs(target, proposal, conditional_distribution, gamma_current)
}
\arguments{
\item{target}{list with keys: \code{logdensity} evaluates log target density, \code{gradlogdensity} returns its gradient}

\item{proposal}{list with keys: \code{rinit} samples from proposal, \code{logdensity} evaluates log proposal density, \code{gradlogdensity} returns its gradient}

\item{conditional_distribution}{the conditional sampling distribution}

\item{gamma_current}{the current inverse temperature}
}
\description{
Run slice Gibbs sampler
}
