% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcsamplers-package.R
\docType{package}
\name{smcsamplers-package}
\alias{smcsamplers-package}
\alias{smcsamplers}
\title{smcsamplers}
\description{
...
}
\details{
...
}
\author{
anonymous <anonymous@anonymous.com>
}
\keyword{package}
