% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperingsmc-package.R
\docType{package}
\name{temperingsmc-package}
\alias{temperingsmc-package}
\alias{temperingsmc}
\title{temperingsmc}
\description{
...
}
\details{
...
}
\author{
?? <anonymous@anonymous.com>
}
\keyword{package}
