% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcsamplers-package.R
\docType{package}
\name{smcsamplers-package}
\alias{smcsamplers-package}
\alias{smcsamplers}
\title{smcsamplers}
\description{
Sequential Monte Carlo samplers
}
\details{
Reproduces the figures in \url{https://arxiv.org/abs/2007.11936}
}
\author{
Chenguang Dai, Jeremy Heng, Pierre E. Jacob <pierre.jacob.work@gmail.com>, Nick Whiteley
}
\keyword{package}
